requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
cryptography==42.0.0

# ML & Deep Learning
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
tensorflow==2.15.0
transformers==4.36.2
huggingface-hub==0.20.2
onnx==1.15.0
onnxruntime==1.16.3

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
resampy==0.4.2

# Video Processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
ffmpeg-python==0.2.0
pillow==10.2.0

# Computer Vision
facenet-pytorch==2.5.3
mediapipe==0.10.9
albumentations==1.3.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4

# Model Compression
tensorflow-model-optimization==0.7.5
neural-compressor==2.4.1

# Cloud Storage
boto3==1.34.22
azure-storage-blob==12.19.0
google-cloud-storage==2.14.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.39.2

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1

# Task Queue
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.0.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
requests==2.31.0
pyyaml==6.0.1
